Refactor Mojo SIMD backend for current Mojo APIs, streamline vector loop lowering, expand benchmarks, and reach near-merge readiness with broad test coverage and significant speedups.

+++ tests/impulse-tests/Make.mojo
@@ -22,7 +22,7 @@ arch ?= mojo/impulse.mojo
 genout ?= archs/mojo
 precision ?=
 FAUSTOPTIONS ?= -I dsp -double
-MOJOBUILDOPTIONS ?= -O3 -D DFAUST=DType.float64
+MOJOBUILDOPTIONS ?= -O3 -D DFAUST=DType.float32
 
 .PHONY: all help test filesCompare clean
 .DELETE_ON_ERROR:
